+++ lib/app/widgets_app.dart
@@ -150,7 +150,7 @@ class BaseWidgetsApp extends StatelessWidget {
               child:
                   builder?.call(context, child) ?? child ?? const SizedBox());
         },
-        navigatorKey: GlobalWayUI().navigatorKey,
+        navigatorKey: FlExtended().navigatorKey,
         routes: routes,
         initialRoute: initialRoute,
         onGenerateRoute: onGenerateRoute,