+++ lib/dependencies/plugin/url_launcher.dart
@@ -83,7 +83,7 @@ class UrlLauncher {
     if ((isIOS || isMacOS) && appId.isNotEmptyOrNull) {
       return await canLaunchUrl(Uri.parse(appId!));
     } else if (isAndroid && packageName.isNotEmptyOrNull) {
-      final appList = await Curiosity().native.getInstalledApps;
+      final appList = await Curiosity.native.getInstalledApps;
       bool installed = false;
       for (var element in appList) {
         if (element.packageName == packageName) {